fix some linting violations
