+++ ios/MullvadVPN/Coordinators/AccountCoordinator.swift
@@ -157,11 +157,13 @@ final class AccountCoordinator: Coordinator, Presentable, Presenting {
             "DEVICE_INFO_DIALOG_MESSAGE_PART_1",
             tableName: "Account",
             value: """
-            This is the name assigned to the device. Each device logged in on a Mullvad account gets a unique name that helps you identify it when you manage your devices in the app or on the website.
+            This is the name assigned to the device. Each device logged in on a Mullvad account gets a unique name \
+            that helps you identify it when you manage your devices in the app or on the website.
 
             You can have up to 5 devices logged in on one Mullvad account.
 
-            If you log out, the device and the device name is removed. When you log back in again, the device will get a new name.
+            If you log out, the device and the device name is removed. When \
+            you log back in again, the device will get a new name.
             """,
             comment: ""
         )
