+++ ios/MullvadVPN/Notifications/Notification Providers/AccountExpiryInAppNotificationProvider.swift
@@ -21,7 +21,7 @@ final class AccountExpiryInAppNotificationProvider: NotificationProvider, InAppN
             didLoadConfiguration: { [weak self] tunnelManager in
                 self?.invalidate(deviceState: tunnelManager.deviceState)
             },
-            didUpdateDeviceState: { [weak self] tunnelManager, deviceState, previousDeviceState in
+            didUpdateDeviceState: { [weak self] _, deviceState, _ in
                 self?.invalidate(deviceState: deviceState)
             }
         )
