+++ ios/MullvadVPN/Presentation controllers/FormsheetPresentationController.swift
@@ -115,7 +115,7 @@ class FormSheetPresentationController: UIPresentationController {
         }
 
         if let transitionCoordinator = presentingViewController.transitionCoordinator {
-            transitionCoordinator.animate { context in
+            transitionCoordinator.animate { _ in
                 revealDimmingView()
             }
         } else {
@@ -137,7 +137,7 @@ class FormSheetPresentationController: UIPresentationController {
         }
 
         if let transitionCoordinator = presentingViewController.transitionCoordinator {
-            transitionCoordinator.animate { context in
+            transitionCoordinator.animate { _ in
                 fadeDimmingView()
             }
         } else {
@@ -167,7 +167,7 @@ class FormSheetPresentationController: UIPresentationController {
         NotificationCenter.default.post(
             name: Self.willChangeFullScreenPresentation,
             object: presentedViewController,
-            userInfo: [Self.isFullScreenUserInfoKey: NSNumber(booleanLiteral: currentIsInFullScreen)]
+            userInfo: [Self.isFullScreenUserInfoKey: NSNumber(value: currentIsInFullScreen)]
         )
     }
