+++ ios/MullvadVPN/View controllers/Account/AccountContentView.swift
@@ -119,438 +119,14 @@ class AccountContentView: UIView {
 
         directionalLayoutMargins = UIMetrics.contentLayoutMargins
 
-        addSubview(contentStackView)
-        addSubview(buttonStackView)
-
-        NSLayoutConstraint.activate([
-            contentStackView.topAnchor.constraint(equalTo: layoutMarginsGuide.topAnchor),
-            contentStackView.leadingAnchor.constraint(equalTo: layoutMarginsGuide.leadingAnchor),
-            contentStackView.trailingAnchor.constraint(equalTo: layoutMarginsGuide.trailingAnchor),
-
+        addConstrainedSubviews([contentStackView, buttonStackView]) {
+            contentStackView.pinEdgesToSuperviewMargins(.all().excluding(.bottom))
             buttonStackView.topAnchor.constraint(
                 greaterThanOrEqualTo: contentStackView.bottomAnchor,
                 constant: UIMetrics.sectionSpacing
-            ),
-            buttonStackView.leadingAnchor.constraint(equalTo: layoutMarginsGuide.leadingAnchor),
-            buttonStackView.trailingAnchor.constraint(equalTo: layoutMarginsGuide.trailingAnchor),
-            buttonStackView.bottomAnchor.constraint(equalTo: layoutMarginsGuide.bottomAnchor),
-        ])
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-class AccountDeviceRow: UIView {
-    var deviceName: String? {
-        didSet {
-            deviceLabel.text = deviceName?.capitalized ?? ""
-            accessibilityValue = deviceName
-        }
-    }
-
-    var infoButtonAction: (() -> Void)?
-
-    private let titleLabel: UILabel = {
-        let label = UILabel()
-        label.text = NSLocalizedString(
-            "DEVICE_NAME",
-            tableName: "Account",
-            value: "Device name",
-            comment: ""
-        )
-        label.font = UIFont.systemFont(ofSize: 14)
-        label.textColor = UIColor(white: 1.0, alpha: 0.6)
-        return label
-    }()
-
-    private let deviceLabel: UILabel = {
-        let label = UILabel()
-        label.font = UIFont.systemFont(ofSize: 17)
-        label.textColor = .white
-        return label
-    }()
-
-    private let infoButton: UIButton = {
-        let button = IncreasedHitButton(type: .system)
-        button.accessibilityIdentifier = "InfoButton"
-        button.tintColor = .white
-        button.setImage(UIImage(named: "IconInfo"), for: .normal)
-        return button
-    }()
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-
-        let contentContainerView = UIStackView(arrangedSubviews: [titleLabel, deviceLabel])
-        contentContainerView.axis = .vertical
-        contentContainerView.alignment = .leading
-        contentContainerView.spacing = 8
-
-        addConstrainedSubviews([contentContainerView, infoButton]) {
-            contentContainerView.pinEdgesToSuperview()
-            infoButton.leadingAnchor.constraint(equalToSystemSpacingAfter: deviceLabel.trailingAnchor, multiplier: 1)
-            infoButton.centerYAnchor.constraint(equalTo: deviceLabel.centerYAnchor)
-        }
-
-        isAccessibilityElement = true
-        accessibilityLabel = titleLabel.text
-
-        infoButton.addTarget(
-            self,
-            action: #selector(didTapInfoButton),
-            for: .touchUpInside
-        )
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    @objc private func didTapInfoButton() {
-        infoButtonAction?()
-    }
-}
-
-class AccountNumberRow: UIView {
-    var accountNumber: String? {
-        didSet {
-            updateView()
-        }
-    }
-
-    var isObscured = true {
-        didSet {
-            updateView()
-        }
-    }
-
-    var copyAccountNumber: (() -> Void)?
-
-    private let titleLabel: UILabel = {
-        let textLabel = UILabel()
-        textLabel.translatesAutoresizingMaskIntoConstraints = false
-        textLabel.text = NSLocalizedString(
-            "ACCOUNT_TOKEN_LABEL",
-            tableName: "Account",
-            value: "Account number",
-            comment: ""
-        )
-        textLabel.font = UIFont.systemFont(ofSize: 14)
-        textLabel.textColor = UIColor(white: 1.0, alpha: 0.6)
-        return textLabel
-    }()
-
-    private let accountNumberLabel: UILabel = {
-        let textLabel = UILabel()
-        textLabel.translatesAutoresizingMaskIntoConstraints = false
-        textLabel.font = UIFont.monospacedSystemFont(ofSize: 17, weight: .regular)
-        textLabel.textColor = .white
-        return textLabel
-    }()
-
-    private let showHideButton: UIButton = {
-        let button = UIButton(type: .system)
-        button.translatesAutoresizingMaskIntoConstraints = false
-        button.tintColor = .white
-        button.setContentHuggingPriority(.defaultHigh, for: .horizontal)
-        return button
-    }()
-
-    private let copyButton: UIButton = {
-        let button = UIButton(type: .system)
-        button.translatesAutoresizingMaskIntoConstraints = false
-        button.tintColor = .white
-        button.setContentHuggingPriority(.defaultHigh, for: .horizontal)
-        return button
-    }()
-
-    private var revertCopyImageWorkItem: DispatchWorkItem?
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-
-        addSubview(titleLabel)
-        addSubview(accountNumberLabel)
-        addSubview(showHideButton)
-        addSubview(copyButton)
-
-        NSLayoutConstraint.activate([
-            titleLabel.topAnchor.constraint(equalTo: topAnchor),
-            titleLabel.leadingAnchor.constraint(equalTo: leadingAnchor),
-            titleLabel.trailingAnchor.constraint(greaterThanOrEqualTo: trailingAnchor),
-
-            accountNumberLabel.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 8),
-            accountNumberLabel.leadingAnchor.constraint(equalTo: leadingAnchor),
-            accountNumberLabel.trailingAnchor.constraint(equalTo: showHideButton.leadingAnchor),
-            accountNumberLabel.bottomAnchor.constraint(equalTo: bottomAnchor),
-
-            showHideButton.heightAnchor.constraint(equalTo: accountNumberLabel.heightAnchor),
-            showHideButton.centerYAnchor.constraint(equalTo: accountNumberLabel.centerYAnchor),
-            showHideButton.leadingAnchor.constraint(equalTo: accountNumberLabel.trailingAnchor),
-
-            copyButton.heightAnchor.constraint(equalTo: accountNumberLabel.heightAnchor),
-            copyButton.centerYAnchor.constraint(equalTo: accountNumberLabel.centerYAnchor),
-            copyButton.leadingAnchor.constraint(
-                equalTo: showHideButton.trailingAnchor,
-                constant: 24
-            ),
-            copyButton.trailingAnchor.constraint(equalTo: trailingAnchor),
-        ])
-
-        showHideButton.addTarget(
-            self,
-            action: #selector(didTapShowHideAccount),
-            for: .touchUpInside
-        )
-
-        copyButton.addTarget(
-            self,
-            action: #selector(didTapCopyAccountNumber),
-            for: .touchUpInside
-        )
-
-        isAccessibilityElement = true
-        accessibilityLabel = titleLabel.text
-
-        showCheckmark(false)
-        updateView()
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    // MARK: - Private
-
-    private func updateView() {
-        accountNumberLabel.text = displayAccountNumber ?? ""
-        showHideButton.setImage(showHideImage, for: .normal)
-
-        accessibilityAttributedValue = _accessibilityAttributedValue
-        accessibilityCustomActions = _accessibilityCustomActions
-    }
-
-    private var displayAccountNumber: String? {
-        guard let accountNumber else {
-            return nil
-        }
-
-        let formattedString = accountNumber.formattedAccountNumber
-
-        if isObscured {
-            return String(formattedString.map { ch in
-                ch == " " ? ch : "•"
-            })
-        } else {
-            return formattedString
-        }
-    }
-
-    private var showHideImage: UIImage? {
-        if isObscured {
-            return UIImage(named: "IconUnobscure")
-        } else {
-            return UIImage(named: "IconObscure")
-        }
-    }
-
-    private var _accessibilityAttributedValue: NSAttributedString? {
-        guard let accountNumber else {
-            return nil
-        }
-
-        if isObscured {
-            return NSAttributedString(
-                string: NSLocalizedString(
-                    "ACCOUNT_ACCESSIBILITY_OBSCURED",
-                    tableName: "Account",
-                    value: "Obscured",
-                    comment: ""
-                )
             )
-        } else {
-            return NSAttributedString(
-                string: accountNumber,
-                attributes: [.accessibilitySpeechSpellOut: true]
-            )
+            buttonStackView.pinEdgesToSuperviewMargins(.all().excluding(.top))
         }
-    }
-
-    private var _accessibilityCustomActions: [UIAccessibilityCustomAction]? {
-        guard accountNumber != nil else { return nil }
-
-        return [
-            UIAccessibilityCustomAction(
-                name: showHideAccessibilityActionName,
-                target: self,
-                selector: #selector(didTapShowHideAccount)
-            ),
-            UIAccessibilityCustomAction(
-                name: NSLocalizedString(
-                    "ACCOUNT_ACCESSIBILITY_COPY_TO_PASTEBOARD",
-                    tableName: "Account",
-                    value: "Copy to pasteboard",
-                    comment: ""
-                ),
-                target: self,
-                selector: #selector(didTapCopyAccountNumber)
-            ),
-        ]
-    }
-
-    private var showHideAccessibilityActionName: String {
-        if isObscured {
-            return NSLocalizedString(
-                "ACCOUNT_ACCESSIBILITY_SHOW_ACCOUNT_NUMBER",
-                tableName: "Account",
-                value: "Show account number",
-                comment: ""
-            )
-        } else {
-            return NSLocalizedString(
-                "ACCOUNT_ACCESSIBILITY_HIDE_ACCOUNT_NUMBER",
-                tableName: "Account",
-                value: "Hide account number",
-                comment: ""
-            )
-        }
-    }
-
-    private func showCheckmark(_ showCheckmark: Bool) {
-        if showCheckmark {
-            let tickIcon = UIImage(named: "IconTick")
-
-            copyButton.setImage(tickIcon, for: .normal)
-            copyButton.tintColor = .successColor
-        } else {
-            let copyIcon = UIImage(named: "IconCopy")
-
-            copyButton.setImage(copyIcon, for: .normal)
-            copyButton.tintColor = .white
-        }
-    }
-
-    // MARK: - Actions
-
-    @objc private func didTapShowHideAccount() {
-        isObscured.toggle()
-        updateView()
-
-        UIAccessibility.post(notification: .layoutChanged, argument: nil)
-    }
-
-    @objc private func didTapCopyAccountNumber() {
-        let delayedWorkItem = DispatchWorkItem { [weak self] in
-            self?.showCheckmark(false)
-        }
-
-        revertCopyImageWorkItem?.cancel()
-        revertCopyImageWorkItem = delayedWorkItem
-
-        showCheckmark(true)
-        copyAccountNumber?()
-
-        DispatchQueue.main.asyncAfter(
-            deadline: .now() + .seconds(2),
-            execute: delayedWorkItem
-        )
-    }
-}
-
-class AccountExpiryRow: UIView {
-    var value: Date? {
-        didSet {
-            let expiry = value
-
-            if let expiry, expiry <= Date() {
-                let localizedString = NSLocalizedString(
-                    "ACCOUNT_OUT_OF_TIME_LABEL",
-                    tableName: "Account",
-                    value: "OUT OF TIME",
-                    comment: ""
-                )
-
-                valueLabel.text = localizedString
-                accessibilityValue = localizedString
-
-                valueLabel.textColor = .dangerColor
-            } else {
-                let formattedDate = expiry.map { date in
-                    DateFormatter.localizedString(
-                        from: date,
-                        dateStyle: .medium,
-                        timeStyle: .short
-                    )
-                }
-
-                valueLabel.text = formattedDate ?? ""
-                accessibilityValue = formattedDate
-
-                valueLabel.textColor = .white
-            }
-        }
-    }
-
-    private let textLabel: UILabel = {
-        let textLabel = UILabel()
-        textLabel.translatesAutoresizingMaskIntoConstraints = false
-        textLabel.text = NSLocalizedString(
-            "ACCOUNT_EXPIRY_LABEL",
-            tableName: "Account",
-            value: "Paid until",
-            comment: ""
-        )
-        textLabel.font = UIFont.systemFont(ofSize: 14)
-        textLabel.textColor = UIColor(white: 1.0, alpha: 0.6)
-        return textLabel
-    }()
-
-    private let valueLabel: UILabel = {
-        let valueLabel = UILabel()
-        valueLabel.translatesAutoresizingMaskIntoConstraints = false
-        valueLabel.font = UIFont.systemFont(ofSize: 17)
-        valueLabel.textColor = .white
-        return valueLabel
-    }()
-
-    let activityIndicator: SpinnerActivityIndicatorView = {
-        let activityIndicator = SpinnerActivityIndicatorView(style: .small)
-        activityIndicator.translatesAutoresizingMaskIntoConstraints = false
-        activityIndicator.tintColor = .white
-        activityIndicator.setContentHuggingPriority(.defaultHigh, for: .horizontal)
-        activityIndicator.setContentCompressionResistancePriority(.defaultHigh, for: .horizontal)
-        return activityIndicator
-    }()
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-
-        addSubview(textLabel)
-        addSubview(activityIndicator)
-        addSubview(valueLabel)
-
-        NSLayoutConstraint.activate([
-            textLabel.topAnchor.constraint(equalTo: topAnchor),
-            textLabel.leadingAnchor.constraint(equalTo: leadingAnchor),
-            textLabel.trailingAnchor.constraint(
-                greaterThanOrEqualTo: activityIndicator.leadingAnchor,
-                constant: -8
-            ),
-
-            activityIndicator.topAnchor.constraint(equalTo: textLabel.topAnchor),
-            activityIndicator.bottomAnchor.constraint(equalTo: textLabel.bottomAnchor),
-            activityIndicator.trailingAnchor.constraint(equalTo: trailingAnchor),
-
-            valueLabel.topAnchor.constraint(equalTo: textLabel.bottomAnchor, constant: 8),
-            valueLabel.leadingAnchor.constraint(equalTo: leadingAnchor),
-            valueLabel.trailingAnchor.constraint(equalTo: trailingAnchor),
-            valueLabel.bottomAnchor.constraint(equalTo: bottomAnchor),
-        ])
-
-        isAccessibilityElement = true
-        accessibilityLabel = textLabel.text
     }
 
     required init?(coder: NSCoder) {
